+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+### 1.0.31 (2018-05-29)
+Bug Fixes:
+* **Snapshot:** Expose the save() method from the snapshot repository
+* **CacheRepository:** The cache is invalidated when the `create()` method is used
+
 ### 1.0.30 (2018-05-24)
 Bug Fixes:
 * **Documentation:** Link to the changelog from the docs